Extract hotkey setup helpers and inject runtime into platform bindings.
Move toggle-command construction into hotkey_setup so GNOME shortcuts,
daemon install, doctor, and tray share one path. Replace Settings
properties with runtime-aware helpers for testability.
